# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ATLAS \u2014 Market Intelligence Engine",
    "description": "Jhaveri Intelligence Platform \u2014 Market \u2192 Sector \u2192 Stock \u2192 Decision",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/stocks/universe": {
      "get": {
        "tags": [
          "stocks"
        ],
        "summary": "Get Universe",
        "description": "Get all stocks grouped by sector with technicals and RS.",
        "operationId": "get_universe_api_v1_stocks_universe_get",
        "parameters": [
          {
            "name": "benchmark",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Benchmark filter",
              "default": "NIFTY 500",
              "title": "Benchmark"
            },
            "description": "Benchmark filter"
          },
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sector filter",
              "title": "Sector"
            },
            "description": "Sector filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockUniverseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stocks/movers": {
      "get": {
        "tags": [
          "stocks"
        ],
        "summary": "Get Movers",
        "description": "Get top RS momentum gainers and losers.",
        "operationId": "get_movers_api_v1_stocks_movers_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 15,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoversResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{symbol}": {
      "get": {
        "tags": [
          "stocks"
        ],
        "summary": "Get Stock Deep Dive",
        "description": "Get complete deep-dive for a stock with conviction pillars.\n\nWhen ``include`` is supplied the modules are validated through the\nshared UQL include layer (:mod:`backend.services.uql.includes`) and\nsurfaced via ``meta.includes_loaded`` per spec \u00a718.2 \u2014 keeping this\nfixed endpoint conformant with the \u00a717/\u00a718/\u00a720 standard. Unknown or\ndeferred modules raise :class:`UQLError(INCLUDE_NOT_AVAILABLE)` which\nthe registered envelope handler converts to a \u00a720.5 400 response.",
        "operationId": "get_stock_deep_dive_api_v1_stocks__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated UQL include modules (spec \u00a718.1) \u2014 e.g. rs,conviction",
              "title": "Include"
            },
            "description": "Comma-separated UQL include modules (spec \u00a718.1) \u2014 e.g. rs,conviction"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockDeepDiveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{symbol}/rs-history": {
      "get": {
        "tags": [
          "stocks"
        ],
        "summary": "Get Rs History",
        "description": "Get RS score history for a stock.",
        "operationId": "get_rs_history_api_v1_stocks__symbol__rs_history_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "benchmark",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "NIFTY 500",
              "title": "Benchmark"
            }
          },
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 60,
              "minimum": 1,
              "default": 12,
              "title": "Months"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RSHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoverEntry": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "rs_composite": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rs Composite"
          },
          "rs_momentum": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rs Momentum"
          },
          "quadrant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Quadrant"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "company_name"
        ],
        "title": "MoverEntry"
      },
      "MoversResponse": {
        "properties": {
          "gainers": {
            "items": {
              "$ref": "#/components/schemas/MoverEntry"
            },
            "type": "array",
            "title": "Gainers"
          },
          "losers": {
            "items": {
              "$ref": "#/components/schemas/MoverEntry"
            },
            "type": "array",
            "title": "Losers"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "gainers",
          "losers",
          "meta"
        ],
        "title": "MoversResponse"
      },
...
      "Quadrant": {
        "type": "string",
        "enum": [
          "LEADING",
          "IMPROVING",
          "WEAKENING",
          "LAGGING"
        ],
        "title": "Quadrant"
      },
      "RSDataPoint": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "rs_composite": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rs Composite"
          },
          "rs_1w": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rs 1W"
          },
          "rs_1m": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rs 1M"
          },
          "rs_3m": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rs 3M"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "RSDataPoint"
      },
...
      "RSHistoryResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "benchmark": {
            "type": "string",
            "title": "Benchmark"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/RSDataPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "benchmark",
          "points",
          "meta"
        ],
        "title": "RSHistoryResponse"
      },
...
      "ResponseMeta": {
        "properties": {
          "data_as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data As Of"
          },
          "record_count": {
            "type": "integer",
            "title": "Record Count",
            "default": 0
          },
          "query_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Ms"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          },
          "returned": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returned"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          },
          "offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "has_more": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has More"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          },
          "cache_hit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Hit"
          },
          "includes_loaded": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Includes Loaded"
          },
          "staleness": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "fresh",
                  "stale",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Staleness"
          }
        },
        "type": "object",
        "title": "ResponseMeta",
        "description": "Response provenance + pagination meta.\n\nV1 fields (data_as_of/record_count/query_ms/stale) remain present and\ntyped identically for backward compatibility (spec \u00a7SC-003). The\nadditional fields below are populated by `services.uql.meta.build_meta`\nfor V2 UQL responses; V1 fixed-endpoint callers leave them as defaults."
      },
...
      "SectorGroup": {
        "properties": {
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "stock_count": {
            "type": "integer",
            "title": "Stock Count"
          },
          "stocks": {
            "items": {
              "$ref": "#/components/schemas/StockSummary"
            },
            "type": "array",
            "title": "Stocks"
          }
        },
        "type": "object",
        "required": [
          "sector",
          "stock_count",
          "stocks"
        ],
        "title": "SectorGroup"
      },
...
      "StockDeepDiveResponse": {
        "properties": {
          "stock": {
            "$ref": "#/components/schemas/StockDeepDive"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "stock",
          "meta"
        ],
        "title": "StockDeepDiveResponse",
        "description": "Single-equity deep-dive response (spec \u00a717.10 + \u00a718.2).\n\nSerializes both ``meta`` (V1 contract \u2014 frontend reads) and ``_meta``\n(the \u00a718 standard path that ``check-api-standard.py`` asserts via\n``uql-04-include-system``). Both keys point at the same payload so a\nresponse is loud about which fields are include-attached without\nbreaking V1 consumers."
      },
...
      "StockUniverseResponse": {
        "properties": {
          "sectors": {
            "items": {
              "$ref": "#/components/schemas/SectorGroup"
            },
            "type": "array",
            "title": "Sectors"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "sectors",
          "meta"
        ],
        "title": "StockUniverseResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}